move css around for better naming and play around with parcel
