--- src/client/theme/SearchModal/SearchModal.module.css
@@ -1,94 +1,0 @@
-.searchModal {
-  position: fixed;
-  width: 100vw;
-  height: 100vh;
-  top: 0;
-  left: 0;
-  background: rgba(94, 100, 112, 0.7);
-  z-index: 100000;
-}
-
-.searchModalContainer {
-  box-shadow: rgba(255, 255, 255, 0.5) 1px 1px 0px 0px inset,
-    rgb(85, 90, 100) 0px 3px 8px 0px;
-  background: var(--search-local-modal-background);
-  position: relative;
-
-  flex-direction: column;
-  margin: 60px auto auto;
-  border-radius: 6px;
-  width: 50%;
-  min-width: 400px;
-  min-height: 200px;
-  border: 1px solid var(--search-local-container-border);
-}
-
-.searchInputContainer {
-  display: flex;
-  margin: 12px;
-  height: 50px;
-  align-items: center;
-  box-shadow: var(--search-local-searchbox-shadow);
-  border-radius: 4px;
-  background-color: var(--search-local-searchbox-focus-background);
-}
-
-.searchInputContainer label {
-  margin-left: 10px;
-  margin-right: 5px;
-  color: var(--search-local-text-color);
-  display: flex;
-}
-
-.searchInputContainer label > svg {
-  stroke-width: 1.6;
-}
-
-.searchInputContainer > span {
-  width: 100%;
-}
-
-.searchInput {
-  padding: 12px;
-  position: relative;
-  width: 100%;
-  height: 100%;
-  color: var(--search-local-text-color);
-  font-size: 1.2em;
-  border: none;
-  outline: none;
-  background-color: transparent;
-}
-
-.messageContainer {
-  width: 100%;
-  height: 150px;
-  text-align: center;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  font-weight: 300;
-  color: var(--search-local-text-color);
-}
-
-.suggestion {
-  margin: 12px;
-}
-
-.searchResultsContainer {
-  overflow-y: scroll;
-  max-height: 400px;
-  color: var(--search-local-text-color);
-}
-
-.loadingContainer {
-  width: 100%;
-  display: flex;
-  place-content: center;
-  margin-top: 25px;
-}
-
-.loadingContainer div {
-  width: 50px;
-  height: 50px;
-}
