--- src/client/theme/SearchModal/SearchResultsSection.module.css
@@ -1,19 +1,0 @@
-.container {
-  overflow-y: scroll;
-  max-height: 400px;
-  color: var(--search-local-text-color);
-}
-
-.header {
-  background: var(--search-local-modal-background);
-  color: var(--search-local-text-color);
-  font-size: 0.85em;
-  font-weight: 600;
-  line-height: 32px;
-  margin: 0 -4px;
-  padding: 8px 4px 0;
-  position: -webkit-sticky;
-  position: sticky;
-  top: 0;
-  z-index: 2;
-}
