+++ src/client/theme/SearchModal/SearchResultsSection.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import styles from "./SearchResultsSection.module.css";
+import styles from "./index.module.css";
 
 interface SearchResultsSectionProps {
   heading: string;
@@ -10,8 +10,8 @@ const SearchResultsSection: React.FC<SearchResultsSectionProps> = (props) => {
   const { children, heading } = props;
 
   return (
-    <section className={styles.container}>
-      <header className={styles.header}>{heading}</header>
+    <section className={styles.searchResultsSection}>
+      <header className={styles.searchResultsSectionHeader}>{heading}</header>
 
       {children}
     </section>
