+++ src/client/theme/SearchModal/index.module.css
@@ -1,0 +1,225 @@
+/* SearchModal Component */
+.searchModal {
+  position: fixed;
+  width: 100vw;
+  height: 100vh;
+  top: 0;
+  left: 0;
+  background: rgba(94, 100, 112, 0.7);
+  z-index: 100000;
+}
+
+.searchModalContainer {
+  box-shadow: rgba(255, 255, 255, 0.5) 1px 1px 0px 0px inset,
+    rgb(85, 90, 100) 0px 3px 8px 0px;
+  background: var(--search-local-modal-background);
+  position: relative;
+
+  flex-direction: column;
+  margin: 60px auto auto;
+  border-radius: 6px;
+  width: 50%;
+  min-width: 400px;
+  min-height: 200px;
+  border: 1px solid var(--search-local-container-border);
+}
+
+.searchInputContainer {
+  display: flex;
+  margin: 12px;
+  height: 50px;
+  align-items: center;
+  box-shadow: var(--search-local-searchbox-shadow);
+  border-radius: 4px;
+  background-color: var(--search-local-searchbox-focus-background);
+}
+
+.searchInputContainer label {
+  margin-left: 10px;
+  margin-right: 5px;
+  color: var(--search-local-text-color);
+  display: flex;
+}
+
+.searchInputContainer label > svg {
+  stroke-width: 1.6;
+}
+
+.searchInputContainer > span {
+  width: 100%;
+}
+
+.searchInput {
+  padding: 12px;
+  position: relative;
+  width: 100%;
+  height: 100%;
+  color: var(--search-local-text-color);
+  font-size: 1.2em;
+  border: none;
+  outline: none;
+  background-color: transparent;
+}
+
+.messageContainer {
+  width: 100%;
+  height: 150px;
+  text-align: center;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  font-weight: 300;
+  color: var(--search-local-text-color);
+}
+
+.suggestion {
+  margin: 12px;
+}
+
+.searchResultsContainer {
+  overflow-y: scroll;
+  max-height: 400px;
+  color: var(--search-local-text-color);
+}
+
+.loadingContainer {
+  width: 100%;
+  display: flex;
+  place-content: center;
+  margin-top: 25px;
+}
+
+.loadingContainer div {
+  width: 50px;
+  height: 50px;
+}
+
+/* SearchResultSection Component */
+.searchResultsSection {
+  overflow-y: scroll;
+  max-height: 400px;
+  color: var(--search-local-text-color);
+}
+
+.searchResultsSectionHeader {
+  background: var(--search-local-modal-background);
+  color: var(--search-local-text-color);
+  font-size: 0.85em;
+  font-weight: 600;
+  line-height: 32px;
+  margin: 0 -4px;
+  padding: 8px 4px 0;
+  position: -webkit-sticky;
+  position: sticky;
+  top: 0;
+  z-index: 2;
+}
+
+/* SearchResult Component */
+.searchResult {
+  cursor: pointer;
+  background: var(--search-local-hit-background, #fff);
+  border-radius: 4px;
+  box-shadow: var(--search-local-hit-shadow, 0 1px 3px 0 #d4d9e1);
+  padding: 0 var(--search-local-spacing, 12px);
+  width: 100%;
+
+  align-items: center;
+  color: var(--search-local-hit-color, #444950);
+  display: flex;
+  flex-direction: row;
+  height: var(--search-local-hit-height, 56px);
+}
+
+html[data-theme="dark"] .searchResult {
+  background: var(--search-local-hit-background, var(--ifm-color-emphasis-100));
+  box-shadow: var(--search-local-hit-shadow, none);
+  color: var(--search-local-hit-color, var(--ifm-font-color-base));
+}
+
+.searchResult:not(:last-child) {
+  margin-bottom: 4px;
+}
+
+.cursor {
+  background-color: var(
+    --search-local-highlight-color,
+    var(--ifm-color-primary)
+  ) !important;
+}
+
+.cursor > span {
+  color: #d4d9e1;
+}
+
+.cursor > .hitWrapper mark {
+  color: #d4d9e1;
+  text-decoration: underline;
+}
+
+.hitTree,
+.hitIcon,
+.hitPath {
+  color: var(--search-local-muted-color, #969faf);
+}
+
+html[data-theme="dark"] .hitTree,
+html[data-theme="dark"] .hitIcon,
+html[data-theme="dark"] .hitPath {
+  color: var(--search-local-muted-color, var(--ifm-color-secondary-darkest));
+}
+
+.hitTree {
+  display: flex;
+  align-items: center;
+}
+
+.hitTree > svg {
+  height: var(--search-local-hit-height, 56px);
+  opacity: 0.5;
+  stroke-width: var(--search-local-icon-stroke-width, 1.4);
+  width: 24px;
+}
+
+.hitIcon {
+  stroke-width: var(--search-local-icon-stroke-width, 1.4);
+
+  height: 20px;
+  width: 20px;
+}
+
+.hitWrapper {
+  flex: 1 1 auto;
+  display: flex;
+  flex-direction: column;
+  font-weight: 500;
+  justify-content: center;
+  margin: 0 8px;
+  overflow-x: hidden;
+  width: 80%;
+}
+
+.hitWrapper mark {
+  background: none;
+  color: var(--search-local-highlight-color, var(--ifm-color-primary));
+}
+
+.hitTitle {
+  font-size: 0.9em;
+}
+
+.hitPath {
+  font-size: 0.75em;
+}
+
+.hitPath,
+.hitTitle {
+  white-space: nowrap;
+  overflow-x: hidden;
+  text-overflow: ellipsis;
+}
+
+.hitAction {
+  height: 20px;
+  width: 20px;
+}
